feat: remove PathingAPI.Container project and add Dockerfile for PathingAPI
