+++ MasterOfPuppets.sln
@@ -36,8 +36,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HeadlessServer", "HeadlessS
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Benchmarks", "Benchmarks\Benchmarks.csproj", "{488984A0-F158-4EDA-A2CB-57C10D8596E1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PathingAPI.Container", "PathingAPI.Container\PathingAPI.Container.csproj", "{96879812-7656-4089-8933-E709F87DB729}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -204,18 +202,6 @@ Global
 		{488984A0-F158-4EDA-A2CB-57C10D8596E1}.Release|x64.Build.0 = Release|Any CPU
 		{488984A0-F158-4EDA-A2CB-57C10D8596E1}.Release|x86.ActiveCfg = Release|Any CPU
 		{488984A0-F158-4EDA-A2CB-57C10D8596E1}.Release|x86.Build.0 = Release|Any CPU
-		{96879812-7656-4089-8933-E709F87DB729}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{96879812-7656-4089-8933-E709F87DB729}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{96879812-7656-4089-8933-E709F87DB729}.Debug|x64.ActiveCfg = Debug|x64
-		{96879812-7656-4089-8933-E709F87DB729}.Debug|x64.Build.0 = Debug|x64
-		{96879812-7656-4089-8933-E709F87DB729}.Debug|x86.ActiveCfg = Debug|x86
-		{96879812-7656-4089-8933-E709F87DB729}.Debug|x86.Build.0 = Debug|x86
-		{96879812-7656-4089-8933-E709F87DB729}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{96879812-7656-4089-8933-E709F87DB729}.Release|Any CPU.Build.0 = Release|Any CPU
-		{96879812-7656-4089-8933-E709F87DB729}.Release|x64.ActiveCfg = Release|x64
-		{96879812-7656-4089-8933-E709F87DB729}.Release|x64.Build.0 = Release|x64
-		{96879812-7656-4089-8933-E709F87DB729}.Release|x86.ActiveCfg = Release|x86
-		{96879812-7656-4089-8933-E709F87DB729}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
