--- PathingAPI.Container/PathingAPI.Container.csproj
@@ -1,32 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <nullable>enable</nullable>
-    <LangVersion>preview</LangVersion>
-    <AssemblyName>PathingAPI.Container</AssemblyName>
-    <RootNamespace>PathingAPI.Container</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\PathingAPI\PathingAPI.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="..\PathingAPI\appsettings.json">
-      <Link>appsettings.json</Link>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </None>
-    <None Include="..\PathingAPI\appsettings.Development.json">
-      <Link>appsettings.Development.json</Link>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </None>
-    <None Include="..\PathingAPI\data_config.json">
-      <Link>data_config.json</Link>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </None>
-  </ItemGroup>
-
-</Project>
